[Feat] 리뷰 작성 기능 + 주석 작성

+++ Week2/Week2/Sources/Model/ReviewModel.swift
@@ -8,7 +8,8 @@
 import Foundation
 
 struct ReviewModel {
-    let starCount: Int?
+    // 별 개수, 제목, 내용 3가지를 받을 간단한 모델을 작성했습니다.
+    let starCount: Double?
     let title: String?
     let contents: String?
 }
